+++ common/include/common/config/GameConfig.h
@@ -44,7 +44,6 @@ struct GameConfig
     CfgVar<Renderer> renderer = Renderer::d3d9;
 
     // Graphics
-    CfgVar<bool> fast_anims = false;
     CfgVar<bool> disable_lod_models = true;
     CfgVar<bool> anisotropic_filtering = true;
     CfgVar<unsigned> msaa = 0;