+++ docs/CHANGELOG.md
@@ -106,6 +106,8 @@ Version 1.1.0 (Tamarack): Not yet released
 - Allow `Anchor_Marker_Orient` event as skybox eye anchor
 - Add support for `$Crater Texture PPM` to `MAPNAME_info.tbl`
 - Add options to `af_ui.tbl` to set the Summoner Trailer button to load a specific level in TC mods
+- Move FastAnimations to `alpine_settings.ini` instead of using RF registry value
+- Add `r_fastanims` console command to toggle fast animations
 
 [@natarii](https://github.com/natarii)
 - Add new damage notification hit sound file